+++ util/opentelemetry-util-genai-evals-deepeval/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this repository are documented in this file.
 
+### Fixed
+- Remove hard-coded temperature argument from LiteLLMModel
+- Support temperature via environment variables for GPT 5 models as those require `TEMPERATURE=1.0`
+
 ## Version 0.1.13 - 2026-02-09
 
 ### Added